[kokoro] Fail when an unknown job is provided.
Prior to this change we were running the bazel build job when an unknown job was provided. This meant that if a job had misconfigured its name it might appear to succeed if the blaze build also happened to succeed.

After this change, the kokoro script will fail if $DEPENDENCY_SYSTEM (the job name) is unknown.