[Typography] Add mdc_scaledFontForTraitEnvironment.
This method's logic exists frequently throughout our componentry and theming extensions as a mechanism for scaling a scalable font with the preferred content size category.

This change introduces a new API that can be invoked by each of the components and theming extensions that support Dynamic Type.

Part of https://github.com/material-components/material-components-ios/issues/7470